Use path parameters for lambdas. SO much cleaner

+++ packages/lambda/src/delete.js
@@ -1,6 +1,22 @@
-module.exports = async function del(event) {
-  const { id } = JSON.parse(event.body);
+const { PrismaClient } = require("@prisma/client");
 
+module.exports = async function del(event) {
+  const { id } = event.pathParameters || {};
+  if (!id) {
+    return {
+      statusCode: 400,
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        error: "Bad Request",
+        workspace: null,
+        output: null,
+      }),
+    };
+  }
+
+  const prisma = new PrismaClient();
   const workspace = await prisma.workspace.delete({
     where: { id },
   });
